Add files via upload
Introduction
This project focuses on finding the highest correlation of gross earning in movie industry.

Correlation :
Correlation refers to a statistical measure that describes the extent to which two variables change together.Correlation does not imply causation; it only shows whether there is a relationship between variables and the direction and strength of that relationship.

There are several types of correlation coefficients used to measure the degree of association between variables:

1. Pearson Correlation Coefficient (Pearson's r):
This is the most common measure of correlation and is used when both variables are continuous and follow a roughly linear relationship. Pearson's r ranges from -1 to 1, with -1 indicating a perfect negative correlation, 1 indicating a perfect positive correlation, and 0 indicating no correlation.

2. Spearman Rank Correlation (Spearman's ρ):
This measure is used when the data is not continuous but ordinal or ranked. It assesses the monotonic relationship between variables. Spearman's ρ also ranges from -1 to 1, with similar interpretations as Pearson's r.

3, Kendall Tau Rank Correlation (Kendall's τ):
Similar to Spearman's ρ, Kendall's τ is used for ranked data, but it measures the strength and direction of the association in terms of concordant and discordant pairs of data points.

We will use Pearson Correlation Coefficient in this project as we have continuous data.

Libraries
The project includes some of the basic Python functions to process, clean, analyze and visualize data.

The Python libraries used are: pandas, seaborn, numpy and matplotlib.